// src/components/global/SearchByBoard/SearchByBoard.enums.ts
export enum CTA_TYPES {
  BRAND = 'brand',
  SIMPLE_SNAP = 'simpleSnap',
  TIRE_SIZE = 'tireSize',
  TYPE = 'type',
  VEHICLE = 'vehicle',
}

export enum CTA_ICONS {
  ICON_BRANDS_CIRCULAR = 'brands-circular',
  ICON_SIMPLE_SNAP = 'camera',
  ICON_TIRE_SIZE = 'wheel',
  ICON_TYPE = 'tire-types-search',
  ICON_VEHICLE = 'car--chrysler-300',
}